requirements.txt
future==0.17.1
gunicorn==19.9.0
pystan==2.18.1.0
Cython==0.29.7
alpha-vantage==2.1.0
fbprophet==0.5
gensim==3.7.3
nltk==3.4.1
pandas==0.24.2
numpy==1.15.4
tweepy==3.7.0
Flask==1.0.3
scikit-learn==0.21.1
requests==2.21.0
Flask-SQLAlchemy==2.4.0
SQLAlchemy==1.3.5
func-timeout==4.3.3
python-dotenv
psycopg2